Events Full information Viewable, options for different events placed

+++ Android/Eternal_Vibes/app/src/main/java/com/example/devandrin/myapplication/Invitation.java
@@ -7,9 +7,9 @@ package com.example.devandrin.myapplication;
 public class Invitation {
     private int id;
     private String Message;
-    private int sender_id ;
+    private long sender_id ;
 
-    public Invitation(int id, String message, int sender_id) {
+    public Invitation(int id, String message, long sender_id) {
         this.id = id;
         Message = message;
         this.sender_id = sender_id;
@@ -31,11 +31,11 @@ public class Invitation {
         Message = message;
     }
 
-    public int getSender_id() {
+    public long getSender_id() {
         return sender_id;
     }
 
-    public void setSender_id(int sender_id) {
+    public void setSender_id(long sender_id) {
         this.sender_id = sender_id;
     }
 }
